GUI and TCP updated
GUI - i have redesign mainUI to its final form (for this moment) and creating showing and hiding mechanic for configuration window widgets.

TCP - i rebuilt the procces of connecting to other player, now player is chosing to host or join game, and after that the app leads his choices according to this decision, for now communication between players is correct, but still i need to make test with 2 different PC.

for now its all, great work for me!